+++ CHANGELOG.rst
@@ -23,7 +23,7 @@ dev
 
 **Bugfixes**
 
--
+- Fix to allow sending 0 bytes on a stream even if the flow control window is negative.
 
 4.3.0 (2025-08-23)
 ------------------